change datatype of a variable

+++ src/main/java/com/qount/invoice/model/Invoice.java
@@ -48,7 +48,7 @@ public class Invoice {
 	private boolean is_mails_automated;
 	private boolean is_cc_current_user;
 	private String payment_spring_customer_id;
-	private double number;
+	private int number;
 	private String po_number;
 	private String document_id;
 	private List<InvoiceLine> invoiceLines;
@@ -326,11 +326,11 @@ public class Invoice {
 		this.payment_spring_customer_id = payment_spring_customer_id;
 	}
 
-	public double getNumber() {
+	public int getNumber() {
 		return number;
 	}
 
-	public void setNumber(double number) {
+	public void setNumber(int number) {
 		this.number = number;
 	}
